requirements.txt
pandas
psycopg2-binary
sqlalchemy
requests
python-dotenv